Refactor interfaces: IHolder and IPlayer
